pyproject.toml

[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
    "requests==2.25.1",
    "beautifulsoup4==4.9.3",
    "m3u8==0.8.0",
    "pycryptodome==3.17",
    "selenium==4.11",
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
